% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-to-eb.r
\name{df_to_eb}
\alias{df_to_eb}
\title{Convert a data frame to a eb_sightings object}
\usage{
df_to_eb(x)
}
\arguments{
\item{x}{data.frame; eBird sightings stored as a data frame.}
}
\value{
An \link{eb_sightings} object.
}
\description{
An \link{eb_sightings} object is just a data frame with the additional class
\code{eb_sightings}. When working with these S3 objects using functions outside
this package, it is possible that the class will be dropped and they will be
reverted to plain data frames. Use \code{df_to_eb} to convert a data frame back to
an \code{eb_sightings} object. Checks are performed to ensure that column names,
ordering, and data types have been preserved.
}
\examples{
# convert sightings to plain data frame
eb_data <- system.file("extdata/MyEBirdData.csv", package = "auklet") \%>\%
  eb_sightings() \%>\%
  as.data.frame()
class(eb_data)
class(df_to_eb(eb_data))
}
