% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb-countable.r
\name{eb_countable}
\alias{eb_countable}
\title{Remove taxa that don't count towards your life list}
\usage{
eb_countable(x)
}
\arguments{
\item{x}{\link{eb_sightings} object; your personal eBird sightings}
}
\value{
An \link{eb_sightings} object.
}
\description{
Removes any records from your eBird data reporting taxa that can't be
resolved to a countable species. This includes slashes, spuhs, undescribed
forms, and domestic species.
}
\examples{
f <- system.file("extdata/MyEBirdData.csv", package = "auklet")
ebird_data <- eb_sightings(f)
nrow(ebird_data)
nrow(eb_countable(ebird_data))
}
