% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb-lifelist.r
\name{eb_lifelist}
\alias{eb_lifelist}
\title{Generate life list}
\usage{
eb_lifelist(x, year, country)
}
\arguments{
\item{x}{\link{eb_sightings} object; your personal eBird sightings}

\item{year}{integer; year or vector of years to calculate list for.}

\item{country}{character; countries to filter by. Countries can either be
expressed as English names or
\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 2-letter country codes}.
English names are matched via regular expressions using
\link[=countrycode]{countrycode}, so there is some flexibility in names.}
}
\value{
A data frame (actually \link[dplyr:tibble]{dplyr::tibble}) of your list list.
}
\description{
Summarize your eBird data into a life list or year list. Lists can be
generated by country.
}
\examples{
f <- system.file("extdata/MyEBirdData.csv", package = "auklet")
ebird_data <- eb_sightings(f)
eb_lifelist(ebird_data)

# 2010 year list for canada
eb_lifelist(ebird_data, year = 2010, country = "CA")
}
