% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb-lifelist-day.r
\name{eb_lifelist_day}
\alias{eb_lifelist_day}
\alias{summary.eb_lifelist_day}
\alias{plot.eb_lifelist_day}
\title{eBird daily life list}
\usage{
eb_lifelist_day(x)

\method{summary}{eb_lifelist_day}(object, by_year = FALSE, ...)

\method{plot}{eb_lifelist_day}(x, title, subtitle, ...)
}
\arguments{
\item{x}{\link{eb_sightings} object; your personal eBird sightings}

\item{object}{\code{eb_lifelist_day} object; your daily life lists.}

\item{by_year}{logical; whether to retain the year first when tallying up
species. Set this to \code{TRUE} if, for example, you're interested in how many
new species you added each year on a given day.}

\item{...}{not used}

\item{title}{character; plot title.}

\item{subtitle}{character; plot subtitle.}
}
\value{
A \link[dplyr:tibble]{dplyr::tibble}, with additional class \code{eb_lifelist_day},
containing your personal sightings.
}
\description{
Generate a daily life list, i.e. for each day of the year list the species
seen on that day regardless of year. For each species, the year in which that
species was first observed is shown. Use \code{\link[=summary]{summary()}} to summarize the list of
species seen each day into counts of the number of species seen each day.
To visualize your day life lists, use \code{\link[=plot]{plot()}}. The concept of the daily
life list was introduced to me by \href{http://www.nemesisbird.com/}{Drew Weber}.
}
\section{Methods (by class)}{
\itemize{
\item \code{eb_lifelist_day}: Summarize daily life lists into counts of species
seen on each day.

\item \code{eb_lifelist_day}: Visualize your day life lists
}}

\examples{
day_list <- system.file("extdata/MyEBirdData.csv", package = "auklet") \%>\%
  eb_sightings() \%>\%
  eb_lifelist_day()
summary(day_list)
plot(day_list)
}
