% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb-sightings.r
\name{eb_sightings}
\alias{eb_sightings}
\title{Read a eBird personal sightings file}
\usage{
eb_sightings(file, countable = FALSE)
}
\arguments{
\item{file}{character; name of your personal eBird sightings files. This file
can be downloaded from the
\href{http://ebird.org/ebird/downloadMyData}{eBird website}.}

\item{countable}{logical; whether to only return countable species, i.e.
those that are counted on your life list by eBird.}
}
\value{
A \link[dplyr:tibble]{dplyr::tibble}, with additional class \code{eb_sightings}, containing
your personal sightings.
}
\description{
Your personal eBird sightings can be downloaded as a CSV file through the
\href{http://ebird.org/ebird/downloadMyData}{eBird website}. This function reads
the eBird data file, cleans up the variable names, and performs some basic
processing.
}
\details{
Some "species" reported on ebird are not true species. For example,
they may be domestic or not resolved to the species level (e.g. \emph{Empidonax
sp.} or \emph{Greater/Lesser Yellowlegs}), in which case they do not count
towards your life list. Alternatively, sightings may be reported below the
species level (e.g. subspecies or recognizable forms), in which case they
must be rolled up to species level prior to reporting. Addiontal fields
from the ebird taxonomy (\code{\link{eb_taxonomy}}) are added to the sightings data
set to resolve these cases. The \code{category} field can be used to identify
different categories of "species". The \code{report_as}, \code{species_common}, and
\code{species_scientific} fields identify the true species corresponding to taxa
reported below species. these fields are all \code{NA} for taxa not resolvable
to species.
}
\examples{
system.file("extdata/MyEBirdData.csv", package = "auklet") \%>\%
  eb_sightings()
}
